Refactor : searchReviews, searchReviewByArea

+++ src/main/java/com/cheffi/review/dto/ReviewPhotoInfoDto.java
@@ -12,13 +12,13 @@ public class ReviewPhotoInfoDto {
     @Schema(description = "리뷰 사진ID", example = "1")
     private Long id;
     @Schema(description = "리뷰에 보여질 사진의 순서", example = "1")
-    private Long photoOrderInReview;
+    private int photoOrderInReview;
     @Schema(description = "리뷰 사진의 URL", example = "https.www.~")
     private String reviewUrl;
 
 
     @Builder
-    public ReviewPhotoInfoDto(Long id, Long photoOrderInReview, String reviewUrl) {
+    public ReviewPhotoInfoDto(Long id, int photoOrderInReview, String reviewUrl) {
         this.id = id;
         this.photoOrderInReview = photoOrderInReview;
         this.reviewUrl = reviewUrl;
